move formatAccompanyingDate into Opportunities helpers
